requirements.txt
python-decouple
swibots